* add jsonwebtoken package. * add utils/auth.js. * implement handling login information.

+++ pages/api/item/delete/[id].js
@@ -1,5 +1,6 @@
 // delete.js
 
+import auth from '@/utils/auth';
 import connectDB from '@/utils/database';
 import { ItemModel } from '@/utils/schemaModels';
 
@@ -7,6 +8,11 @@ const deleteItem = async (req, res) => {
   try {
     const id = req.query.id;
     await connectDB();
+    const singleItem = await ItemModel.findById(id);
+    const loginEmail = req.body.email;
+    if (singleItem.email !== loginEmail) {
+      throw new Error();
+    }
     await ItemModel.deleteOne({ _id: id });
     return res.status(200).json({ message: 'アイテム削除成功' });
   } catch (err) {
@@ -14,4 +20,4 @@ const deleteItem = async (req, res) => {
   }
 };
 
-export default deleteItem;
+export default auth(deleteItem);
